LPS-97666 | LPS-103406 SF

+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/src/main/java/com/liferay/portal/workflow/metrics/rest/internal/resource/v1_0/MetricResourceImpl.java
@@ -281,7 +281,7 @@ public class MetricResourceImpl extends BaseMetricResourceImpl {
 			timeAmount = timeRange / 365.0;
 		}
 
-		return (double)bucket.getDocCount() / timeAmount;
+		return bucket.getDocCount() / timeAmount;
 	}
 
 	@Reference
